finally set up the initial ORM configuration with Sequelize

+++ backend/Controller/FavoritedItemsController.js
@@ -1,22 +1,26 @@
 const { where } = require('sequelize')
 const Customer = require('../Model/Customer')
 const FavoritedItem = require('../Model/FavoritedItem')
-const Shoe = require('../Model/Shoe')
 
 const getFavoriteItems = async (req, res) => {
     try {
         const { email } = req.params
 
-        const results = await Shoe.findAll({
-            include: [{
-                model: FavoritedItem,
-                where: { cus_email: email},
-                attributes: []
-            }],
-            attributes: ['id', 'name', 'star', 'price', 'img'],
+        const cus_id = await Customer.findOne({
+            where: { email: email },
+            attributes: ['id']
+        })
+
+        if (!cus_id) {
+            return res.status(200).json({ success: false, message: "User not found." })
+        }
+
+        const results = await FavoritedItem.findAll({
+            where: { cus_id: cus_id.id},
+            attributes: ['shoe_id'],
             raw: true
         })
-        console.log(results)
+
         if (results.length > 0) {
             return res.status(200).json({ success: true, data: results })
         }
@@ -31,15 +35,24 @@ const getFavoriteItems = async (req, res) => {
 
 const deleteFavoriteItem = async (req, res) => {
     try {
-        const { email, shoe } = req.query
+        const { email, shoe_id } = req.query
+
+        const cus_id = await Customer.findOne({
+            where: { email: email },
+            attributes: ['id']
+        })
+
+        if (!cus_id) {
+            return res.status(200).json({ success: false, message: "User not found." })
+        }
 
         const result = await FavoritedItem.destroy({
             where: {
-                cus_email: email,
-                shoe_name: shoe
+                cus_id: cus_id.id,
+                shoe_id: shoe_id
             }
         })
-        console.log(result)
+
         if (result) {
             return res.status(200).json({ success: true, message: "Delete successfully."})
         }
@@ -53,7 +66,7 @@ const deleteFavoriteItem = async (req, res) => {
 
 const postFavoriteItem = async (req, res) => {
     try {
-        const { email, shoe } = req.query
+        const { email, shoe_id } = req.query
 
         const cus_id = await Customer.findOne({
             where: { email: email },
@@ -66,7 +79,7 @@ const postFavoriteItem = async (req, res) => {
 
         await FavoritedItem.create({
             cus_id: cus_id.id,
-            shoe_id: shoe
+            shoe_id: shoe_id
         })
 
         return res.status(200).json({ success: true, message: "Data inserted successfully." })
